Refactor duplicated utility function into utils
