+++ app/utils.js
@@ -1,0 +1,21 @@
+// To make it easier to use in the view, format express-validator errors.
+function formatValidationErrors (errorsInstance) {
+  if (errorsInstance.isEmpty()) {
+    return false
+  }
+  const errors = errorsInstance.array()
+  const formattedErrors = {}
+  errors.forEach(error => {
+    formattedErrors[error.param] = {
+      id: error.param,
+      href: '#' + error.param,
+      value: error.value,
+      text: error.msg
+    }
+  })
+  return formattedErrors
+}
+
+module.exports = {
+  formatValidationErrors
+}
